third_party/WebKit/LayoutTests/http/tests/security/secureContexts/127.0.0.1.html

<!DOCTYPE html>
<html>
<head>
    <title>127.0.0.1 is secure</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/get-host-info.js"></script>
</head>
<body>
    <script>
        test(function () {
            assert_equals(document.location.hostname, "127.0.0.1", "Sanity check the test runner.");
            assert_true(window.isSecureContext);
        }, "127.0.0.1 is secure.");
    </script>
</body>
</html>
