third_party/WebKit/LayoutTests/platform/win/http/tests/security/secureContexts/127.0.0.1-expected.txt

This is a testharness.js-based test.
PASS 127.0.0.1 is secure. 
Harness: the test ran to completion.
